refactor: MySQLTestConfig를 공용 테스트 패키지로 이동
도메인과 무관한 테스트 인프라 설정이므로
com.realteeth.assignment.domain에서 상위 패키지로 이동한다.
