+++ src/test/java/com/realteeth/assignment/domain/JobRepositoryTest.java
@@ -11,6 +11,7 @@ import org.springframework.boot.data.jpa.test.autoconfigure.DataJpaTest;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.boot.jdbc.test.autoconfigure.AutoConfigureTestDatabase;
+import com.realteeth.assignment.MySQLTestConfig;
 import org.springframework.context.annotation.Import;
 
 import static org.assertj.core.api.Assertions.assertThat;
